refactor: rename mixins.py -> _mixins.py; export CLIArg from top-level package
Both mixins.py and _output.py are internal implementation modules whose
public symbols are re-exported via __init__.py. Aligning the naming by
making mixins private too.

CLIArg was the only symbol requiring a direct import from microbench.mixins
(used by custom mixin authors). It is now exported from the top-level
package so users can write `from microbench import CLIArg`. _UNSET remains
private; the docs example is updated to import it from microbench._mixins.
